rename main.go to handler.go
