+++ handler.go
@@ -1,0 +1,30 @@
+package main
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+	"os"
+)
+
+func helloHandler(w http.ResponseWriter, r *http.Request) {
+	message := "This HTTP triggered function executed successfully. Pass a name in the query string for a personalized response.\n"
+	name := r.URL.Query().Get("name")
+	if name != "" {
+		w.Write([]byte("This HTTP trigger function executed successfully."))
+		//message = fmt.Sprintf("Hello, %s. This HTTP triggered function executed successfully.\n", name)
+	} else {
+		w.Write([]byte(fmt.Sprintf("Hello, %s. This HTTP triggered function executed successfully.", name)))
+	}
+	fmt.Fprint(w, message)
+}
+
+func main() {
+	listenAddr := ":8080"
+	if val, ok := os.LookupEnv("FUNCTIONS_CUSTOMHANDLER_PORT"); ok {
+		listenAddr = ":" + val
+	}
+	http.HandleFunc("/api/HttpExample", helloHandler)
+	log.Printf("About to listen on %s. Go to https://127.0.0.1%s/", listenAddr, listenAddr)
+	log.Fatal(http.ListenAndServe(listenAddr, nil))
+}
